Refactor: Extract IDE code transformation logic to separate IdeCodeTransformer class
